fix reward finalized
